# api/swagger.yaml
basePath: null
definitions: {}
info:
    description: 'A Typescript framework for home automation'
    license:
        name: GPL-3.0
    title: domoja
    version: 1.2.0
paths:
    /devices:
        get:
            operationId: DevicesServiceGetDevices
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Retrieves the list of devices'
            parameters: []
    '/devices/{id}':
        get:
            operationId: DevicesServiceGet
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Retrieves a device'
            parameters:
                -
                    description: 'path of the device'
                    in: path
                    name: id
                    required: true
                    type: string
        post:
            operationId: DevicesServiceSendCommand
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Sends a command to a device'
            consumes:
                - application/x-www-form-urlencoded
            parameters:
                -
                    description: 'path of the device'
                    in: path
                    name: id
                    required: true
                    type: string
                -
                    description: 'command sent to the device'
                    in: formData
                    name: command
                    required: true
                    type: string
    '/devices/{id}/snapshot':
        get:
            operationId: DevicesServiceGetSnapshot
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Get a snapshot from a camera device'
            parameters:
                -
                    description: 'path of the device'
                    in: path
                    name: id
                    required: true
                    type: string
    /pages:
        get:
            operationId: PagesServiceGetPages
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Retrieves the list of pages'
            parameters: []
    /app:
        get:
            operationId: AppServiceGetApp
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Retrieve the app data'
            parameters: []
    /app/demo-mode:
        post:
            operationId: AppServiceSetDemoMode
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: 'Set the app demo mode'
            consumes:
                - application/x-www-form-urlencoded
            parameters:
                -
                    description: ""
                    in: formData
                    name: value
                    required: true
                    type: boolean
swagger: '2.0'
securityDefinitions: {}